Revise Python indent behaviour.
This aims to follow VS Code's behaviour and biases towards not changing
the indent as the user may be in the middle of rearranging code.

As an exception, we implement dedent behaviour for return and raise.
VS Code doesn't do this as they don't have a way to tell if you're in
the middle of the expression after the return or raise.

Significant changes:

- Adds dedent behaviour for constructs that must end a body.
- Adds delimited indents for argument lists, parameter lists and
  parathesized expressions. The latter is necessary for tuple indents to
  work properly as a tuple isn't parsed as such when first being written.
- Doesn't fight manual dedents. Previously trying to dedent then leave a
  blank line or two before the next statement would result in a reindent.
- Removes dedent for else and similar to match VS Code. VS Code has
  attempted to add these in the past but has had to revert because of
  users finding them more annoying than helpful. The implementation here
  was double-dedenting if the user dedented manually and wrote a valid
  else statement.